+++ Chapter08/MySQL/Maven/Basic/OneToMany/src/main/java/com/classicmodels/pojo/SimpleUProduct.java
@@ -9,9 +9,9 @@ public class SimpleUProduct implements Serializable {
     
     private String productName;
     private String productVendor;
-    private Short quantityInStock;
+    private Integer quantityInStock;
 
-    public SimpleUProduct(String productName, String productVendor, Short quantityInStock) {
+    public SimpleUProduct(String productName, String productVendor, Integer quantityInStock) {
         this.productName = productName;
         this.productVendor = productVendor;
         this.quantityInStock = quantityInStock;
@@ -33,11 +33,11 @@ public class SimpleUProduct implements Serializable {
         this.productVendor = productVendor;
     }
 
-    public Short getQuantityInStock() {
+    public Integer getQuantityInStock() {
         return quantityInStock;
     }
 
-    public void setQuantityInStock(Short quantityInStock) {
+    public void setQuantityInStock(Integer quantityInStock) {
         this.quantityInStock = quantityInStock;
     }
